Fixed error when using custom dataset
For a custom dataset, it's most likely the images to be named as `string` like `a1.jpg`.
The `if _isArrayLike(ids):` line will makes this error since python see a string as an array.
The `self.imgs` is a `dict` so there's absolutely no reason not to use a `string` as `ids`. 
This edit won't affect the official `int` ids.