process instances with same score in one step
When computing points on the precision - recall curve instances with
the same score should be processed in one step, not sequentially.
If they have the same score there is no fixed order, but if they are
processed sequentially the result depends on the order. In many typical
use-cases this might not significant, however the fewer instances there
are the higher the impact.